adding bits

+++ Contest#1/326175_F.cpp
@@ -8,7 +8,7 @@ int main()
     std::string bin_a = std::bitset<32>(a).to_string();
     std::string bin_b = std::bitset<32>(b).to_string();
     std::string bin_bitwise;
-    for (u_int8_t i = 0; i < 32; i++)
+    for (short i = 0; i < 32; i++)
     {
         if ((bin_a[i] == '1' && bin_b[i] == '1') || (bin_a[i] == '0' && bin_b[i] == '0'))
             bin_bitwise += '0';
